+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Refreshed modal dialog styling with subtle animations
 - Primary action button moved to the right side of dialog
+- Introduced custom analyticsPrimary button style and updated dialog buttons
 
 ## 5.6.2 - 2025-06-10
 ### Fixed